Make Zoom One Type updatable.
Fixes #2

Not sure if this works on creation or not, but update and read work. Not
sure if create worked as it was coded. This will wipe out phone feature
on create as coded. This is what NDSU wants, as we require a different
API call before phone feature can be turned on anyway.

Also change type over to long, as our Zoom One Type is way longer than
an integer in Java.

+++ src/main/java/com/exclamationlabs/connid/base/zoom/model/ZoomFeature.java
@@ -4,12 +4,12 @@ import com.google.gson.annotations.SerializedName;
 
 public class ZoomFeature {
   @SerializedName("zoom_one_type")
-  private Integer zoomOneType;
+  private Long zoomOneType;
 
   @SerializedName("zoom_phone")
   private Boolean zoomPhone;
 
-  public Integer getZoomOneType() {
+  public Long getZoomOneType() {
     return zoomOneType;
   }
 
@@ -17,7 +17,7 @@ public class ZoomFeature {
     return zoomPhone;
   }
 
-  public void setZoomOneType(Integer zoomOneType) {
+  public void setZoomOneType(Long zoomOneType) {
     this.zoomOneType = zoomOneType;
   }
